move pkg/client/unversioned/record to pkg/client/record; it's version agnostic
